Add compressed variants of the FEED and TEMP_FEED commands (client).
This adds alternative commands for the feed and temp_feed that will
eventually support compression.  This does not implement the protocol
itself, but instead the auto-discovery of the availability of the
protocol and its preferred use.

Stratcon -> Noit connections now do a CAPA command before establishing
a feed connection.  If the desired upgraded protocol is avialable,
we attempt to use that, otherwise we use the base protocol.